+++ RELEASE_NOTES.md
@@ -7,6 +7,11 @@ As such, to be able to move swiftly (pun intended) and not get stuck with a bunc
 In `v4.0`, this library will become a part of the main library. After that, semver will apply to the SwiftUI parts as well.  
 
 
+## 3.4.1
+
+This release fixes some visual artefacts in the callout bubbles.
+
+
 ## 3.4.0
 
 This release adds support for input callouts and secondary input callouts.